feat: add button story for usage documentation Remove example story
